api/requirements.txt
python-dotenv
sqlalchemy
loguru
python-jose
aiosqlite
fastapi
uvicorn
passlib
aiohttp